Move BoxViewDetector and SummaryScreenDetector to their own files.
