+++ SerialPrograms/Source/PokemonHome/Inference/PokemonHome_ButtonDetector.cpp
@@ -156,34 +156,6 @@ bool ButtonDetector::detect(const ImageViewRGB32& screen){
 }
 
 
-BoxViewDetector::BoxViewDetector(VideoOverlay* overlay) : m_button_plus_detector(COLOR_BLACK, ButtonType::ButtonPlus, {0.100, 0.956, 0.107, 0.041}, overlay){}
-
-void BoxViewDetector::make_overlays(VideoOverlaySet& items) const{
-    m_button_plus_detector.make_overlays(items);
-}
-
-bool BoxViewDetector::detect(const ImageViewRGB32& screen){
-    return m_button_plus_detector.detect(screen);
-}
-
-
-SummaryScreenDetector::SummaryScreenDetector(VideoOverlay* overlay) : m_button_B_detector(COLOR_BLACK, ButtonType::ButtonB, {0.100, 0.956, 0.107, 0.041}, overlay){}
-
-void SummaryScreenDetector::make_overlays(VideoOverlaySet& items) const{
-    m_button_B_detector.make_overlays(items);
-}
-
-bool SummaryScreenDetector::detect(const ImageViewRGB32& screen){
-    return m_button_B_detector.detect(screen);
-}
-
-
-
-
-
-
-
-
 }
 }
 }
