+++ SerialPrograms/Source/PokemonHome/Programs/PokemonHome_BoxSorter.cpp
@@ -33,13 +33,16 @@ language
 #include "CommonFramework/ImageTools/ImageStats.h"
 #include "CommonFramework/Notifications/ProgramNotifications.h"
 #include "CommonFramework/ProgramStats/StatsTracking.h"
+#include "CommonFramework/VideoPipeline/VideoOverlay.h"
+#include "CommonFramework/VideoPipeline/VideoOverlayScopes.h"
 #include "CommonTools/Async/InferenceRoutines.h"
 #include "CommonTools/StartupChecks/StartProgramChecks.h"
 #include "NintendoSwitch/Commands/NintendoSwitch_Commands_PushButtons.h"
 #include "Pokemon/Pokemon_Strings.h"
 #include "Pokemon/Pokemon_BoxCursor.h"
 #include "Pokemon/Pokemon_CollectedPokemonInfo.h"
-#include "PokemonHome/Inference/PokemonHome_ButtonDetector.h"
+#include "PokemonHome/Inference/PokemonHome_BoxViewDetector.h"
+#include "PokemonHome/Inference/PokemonHome_SummaryScreenDetector.h"
 #include "PokemonHome_BoxNavigation.h"
 #include "PokemonHome_BoxSorter.h"
